[Refactor]: Merge layer wise lr decay to Default constructor
