+++ mmselfsup/core/optimizer/__init__.py
@@ -1,11 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .builder import build_optimizer
 from .constructor import DefaultOptimizerConstructor
-from .mae_finetune_constructor import MAEFtOptimizerConstructor
 from .optimizers import LARS
-from .simmim_constructor import SimMIMFtOptimizerConstructor
 
-__all__ = [
-    'LARS', 'build_optimizer', 'DefaultOptimizerConstructor',
-    'MAEFtOptimizerConstructor', 'SimMIMFtOptimizerConstructor'
-]
+__all__ = ['LARS', 'build_optimizer', 'DefaultOptimizerConstructor']
